# openapi.yaml
openapi: 3.0.3
info:
  title: P0rtal RDP Broker API
  description: API for managing RDP sessions through the P0rtal broker proxy.
  version: 3.0.0
servers:
  - url: http://localhost:8080
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >-
        API key passed as Bearer token. Used for management API endpoints
        (access control, public keys, recordings, secrets).
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Signed JWT for user access API (POST /api/sessions). The JWT must
        contain a `sub` claim (user email) and be signed with a private key
        whose public key is registered via /api/public-keys.
  schemas:
...
    CreateSessionRequest:
      type: object
      required: [hostname]
      properties:
        hostname:
          type: string
          description: Hostname of the target machine to connect to
          example: mike-rdp
    CreateSessionResponse:
      type: object
      properties:
        session_id:
          type: string
        proxy_host:
          type: string
        proxy_port:
          type: integer
        rdp_file_content:
          type: string
          description: Full RDP file content. Save as .rdp and open with an RDP client.
        state:
          type: string
          enum: [creating, active, connected, terminating, terminated]
        created_at:
          type: string
          format: date-time
        token_expires_at:
          type: string
          format: date-time
    Session:
      type: object
      properties:
        session_id:
          type: string
        user_id:
          type: string
        target_id:
          type: string
          description: Hostname of the target
        target_host:
          type: string
        proxy_port:
          type: integer
        state:
          type: string
          enum: [creating, active, connected, terminating, terminated]
        pid:
          type: integer
        created_at:
          type: string
          format: date-time
        connected_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    CreateRecordingRequest:
      type: object
      properties:
        session_id:
          type: string
        target_id:
          type: string
        target_name:
          type: string
        windows_user:
          type: string
        proxy_user:
          type: string
        agent_hostname:
          type: string
        chunk_secs:
          type: integer
    Recording:
      type: object
      properties:
        id:
          type: string
        session_id:
          type: string
        target_id:
          type: string
        target_name:
          type: string
        windows_user:
          type: string
        proxy_user:
          type: string
        agent_hostname:
          type: string
        status:
          type: string
          enum: [recording, completed, failed]
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
          nullable: true
        duration_secs:
          type: number
        chunk_secs:
          type: integer
        chunk_count:
          type: integer
        total_bytes:
          type: integer
        event_count:
          type: integer
    RecordingEvent:
      type: object
      properties:
        ts:
          type: string
          format: date-time
        type:
          type: string
          enum: [process_start, process_end, window_focus, mouse_click, clipboard, winlog]
        data:
          type: object
          additionalProperties: true
paths:
...
  /api/sessions:
    post:
      summary: Create RDP session
      description: >-
        Creates a new RDP session. Requires a signed JWT (not an API key).
        The user's identity is extracted from the JWT `sub` claim. The proxy
        finds the matching grant by (email, hostname) and uses the stored
        host + user config to create the session.
      tags: [Sessions]
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSessionResponse"
        "400":
          description: Missing hostname
        "401":
          description: Invalid or missing JWT
        "403":
          description: No matching grant for this user + hostname
        "503":
          description: Session limit reached or secret resolution failed
    get:
      summary: List sessions
      tags: [Sessions]
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
...
  /api/recordings:
    post:
      summary: Create recording
      tags: [Recordings]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordingRequest"
      responses:
        "201":
          description: Recording created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
    get:
      summary: List recordings
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: target
          in: query
          schema:
            type: string
        - name: user
          in: query
          schema:
            type: string
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
        - name: status
          in: query
          schema:
            type: string
            enum: [recording, completed, failed]
      responses:
        "200":
          description: List of recordings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recording"
  /api/recordings/{id}:
    get:
      summary: Get recording metadata
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recording metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Recording not found
    delete:
      summary: Delete recording
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Recording deleted
        "404":
          description: Recording not found
  /api/recordings/{id}/chunks:
    post:
      summary: Upload video chunk
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          video/mp2t:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Chunk uploaded
        "404":
          description: Recording not found
  /api/recordings/{id}/events:
    post:
      summary: Send recording events
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/RecordingEvent"
      responses:
        "204":
          description: Events received
        "404":
          description: Recording not found
    get:
      summary: Get recording events
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecordingEvent"
        "404":
          description: Recording not found
  /api/recordings/{id}/end:
    put:
      summary: End recording
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Recording finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recording"
        "404":
          description: Recording not found
  /api/recordings/{id}/video:
    get:
      summary: Stream recording video
      tags: [Recordings]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Video file
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "404":
          description: Recording or video not found